feat(http): prepare new additions.
